Title: n개의 정수 중 최대 2개, Time: 49ms, Memory: 0MB, Status: Runtime Error - CodeTree

+++ 250127/n개의 정수 중 최대 2개/two-max-of-n-num.cpp
@@ -9,7 +9,7 @@ int main() {
     // Write your code here!
     int n;
     cin>>n;
-    vector<int> v;
+    vector<long long> v;
     for(int i=0; i<n; i++){
         int a;
         cin>>a;
